requirements.txt
streamlit>=1.38.0
torch>=2.0.1
torchvision>=0.15.2
pillow>=9.5.0
pandas>=2.1.0
python-dotenv>=1.0.0
huggingface_hub>=0.15.1
langchain-openai>=0.1.0
langchain-groq>=0.1.3
faiss-cpu>=1.7.4
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.2
numpy>=1.24.4
requests>=2.31.0
scikit-learn>=1.3.0
python-magic>=0.4.27
